requirements.txt
streamlit
numpy==1.23.5
pandas
pyradiomics
nibabel
opencv-python
scikit-learn